+++ apps/explorer/lib/explorer/helper.ex
@@ -92,4 +92,17 @@ defmodule Explorer.Helper do
       decoded -> decoded
     end
   end
+
+  @doc """
+  Checks if input is a valid URL
+  """
+  @spec validate_url(String.t() | nil) :: {:ok, String.t()} | :error
+  def validate_url(url) when is_binary(url) do
+    case URI.parse(url) do
+      %URI{host: nil} -> :error
+      _ -> {:ok, url}
+    end
+  end
+
+  def validate_url(_), do: :error
 end